[BE] refactor : 미팅 DTO 패키지 수정
