+++ backend/src/main/java/moim_today/dto/meeting/meeting/MeetingCreateRequest.java
@@ -1,4 +1,4 @@
-package moim_today.dto.meeting;
+package moim_today.dto.meeting.meeting;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
