+++ backend/src/test/java/moim_today/fake_class/meeting/meeting/FakeMeetingService.java
@@ -2,13 +2,12 @@ package moim_today.fake_class.meeting.meeting;
 
 import moim_today.application.meeting.meeting.MeetingService;
 import moim_today.domain.meeting.enums.MeetingStatus;
-import moim_today.dto.meeting.MeetingCreateRequest;
 import moim_today.dto.meeting.MeetingCreateResponse;
-import moim_today.dto.meeting.MeetingDetailResponse;
-import moim_today.dto.meeting.MeetingSimpleResponse;
+import moim_today.dto.meeting.meeting.MeetingCreateRequest;
+import moim_today.dto.meeting.meeting.MeetingDetailResponse;
+import moim_today.dto.meeting.meeting.MeetingSimpleResponse;
 import moim_today.dto.member.MemberSimpleResponse;
 import moim_today.global.error.ForbiddenException;
-import moim_today.util.TestConstant;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
