+++ backend/src/test/java/moim_today/implement/moim/moim/MoimFinderTest.java
@@ -20,7 +20,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
@@ -257,7 +256,8 @@ class MoimFinderTest extends ImplementTest {
                 .build();
 
         moimRepository.save(firstCreatedMoimJpaEntity);
-        moimRepository.flush();
+
+        Thread.sleep(10);
 
         Thread.sleep(10);
 
@@ -267,7 +267,6 @@ class MoimFinderTest extends ImplementTest {
                 .build();
 
         moimRepository.save(secondCreatedMoimJpaEntity);
-        moimRepository.flush();
 
         MoimCategoryDto moimCategoryDto = MoimCategoryDto.ALL;
         MoimSortedFilter moimSortedFilter = null;
@@ -291,7 +290,8 @@ class MoimFinderTest extends ImplementTest {
                 .build();
 
         moimRepository.save(firstCreatedMoimJpaEntity);
-        moimRepository.flush();
+
+        Thread.sleep(10);
 
         Thread.sleep(10);
 
@@ -301,7 +301,6 @@ class MoimFinderTest extends ImplementTest {
                 .build();
 
         moimRepository.save(secondCreatedMoimJpaEntity);
-        moimRepository.flush();
 
         MoimCategoryDto moimCategoryDto = MoimCategoryDto.ALL;
         MoimSortedFilter moimSortedFilter = MoimSortedFilter.CREATED_AT;
